requirements.txt
# Core Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1
django-rest-auth==0.9.5

# File Handling
Pillow
django-storages==1.14.2

# API Documentation for the application
drf-yasg==1.21.7

# Utilities
django-filter==23.5
django-environ==0.11.2

# WSGI Server
gunicorn==21.2.0
whitenoise==6.6.0